feat: disable color in policy report when writing to file
refactor: move policy report methods to output/policy.go
